added rooms class and started work on implementing room colliders. colliders are successfully generated for the walls using room_expected.txt, but cannot get door colliders to generate

+++ src/Collision.h
@@ -15,7 +15,7 @@ struct RaycastData{
 };
 
 struct SortingData{
-	Uint64 index;
+	int index;
 	double contactTime;
 };
